backend/requirements.txt
python-dotenv
faiss-cpu==1.10.0
pandas
numpy
flask 
flask-cors
requests
gunicorn
